mgr/cephadm: move _remote_connection to serve.py
